代码梳理与优化：布局的ChildMargin变量类型从int32_t变成uint16_t，用uint16_t足够表示ChildMargin的范围，可用节约少许空间。

+++ duilib/Box/Layout.h
@@ -145,10 +145,10 @@ private:
 	UiPadding m_rcPadding;
 
 	//子控件之间的额外边距: X 轴方向
-	int32_t m_iChildMarginX;
+	uint16_t m_iChildMarginX;
 
 	//子控件之间的额外边距: Y 轴方向
-	int32_t m_iChildMarginY;
+	uint16_t m_iChildMarginY;
 };
 
 } // namespace ui
